Rework flow of routes to match new design.
We drop some concepts. Activity is now "Deposits" since that is what we
deal with. There is no longer a rigid process of upload -> select DOI
-> confirm citations -> deposit. After an upload, the user works out of
a "/deposit" route, with the ability to manipulate each section of the
deposit displayed.

There is no longer a "/doi" route - DOI is selected either before upload
or inline on the deposit information page ("/deposit/:name").

Deposit status - that is, the status of the deposit with the main CrossRef
system, is moved to the main deposit information page. Depositing and
redepositing can be controlled from this section.